Fixed issue with the default Endpoint Components being declared in a separate file
