--- lib/components/WithEndpoint/Defaults.tsx
@@ -1,11 +1,0 @@
-import { WithEndpoint } from ".";
-
-import { Form, Button, DropdownButton } from "react-bootstrap";
-
-const EndpointInput = WithEndpoint(Form.Control);
-const EndpointButton = WithEndpoint(Button);
-const EndpointDropdown = WithEndpoint(DropdownButton);
-const EndpointCheckbox = WithEndpoint(Form.Check);
-
-
-export {EndpointInput, EndpointButton, EndpointDropdown, EndpointCheckbox};
