+++ lib/main.ts
@@ -1,6 +1,5 @@
 export { useAdapterEndpoint } from './components/AdapterEndpoint';
-export { WithEndpoint } from './components/WithEndpoint';
-export {EndpointInput, EndpointButton, EndpointDropdown, EndpointCheckbox } from './components/WithEndpoint/Defaults';
+export { WithEndpoint, EndpointInput, EndpointButton, EndpointDropdown, EndpointCheckbox } from './components/WithEndpoint';
 export { TitleCard } from './components/TitleCard';
 export { OdinApp } from './components/OdinApp';
 export { OdinGraph } from './components/OdinGraph';
